refactor(authz): separate authz structs in permission, role and context files
